spring08 add test

+++ spring-08/src/main/java/ru/otus/spring/spring08/service/SequenceGeneratorService.java
@@ -23,12 +23,12 @@ public class SequenceGeneratorService {
         this.mongoOperations = mongoOperations;
     }
 
-    public long generateSequence(String seqName) {
+    public int generateSequence(String seqName) {
 
         DatabaseSequence counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
                 new Update().inc("seq",1), options().returnNew(true).upsert(true),
                 DatabaseSequence.class);
-        return !Objects.isNull(counter) ? counter.getSeq() : 1;
+        return !Objects.isNull(counter) ? (int) counter.getSeq() : 1;
 
     }
 }
